refactor: test용 mock함수의 파일 명 변경
- mocks/test.ts 시 테스트 코드로 인식되어 테스트 되는 경우가 있어 mockTestFunction.ts로 변경하였음.
